// gen/tenant/v1/profile.schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/v1/tenant/{tenantId}/profile": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get: operations["ProfileService_GetProfile"];
    put: operations["ProfileService_UpsertProfile"];
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
    ProfileServiceUpsertProfileBody: {
      postalCode?: string;
      prefecture?: string;
      city?: string;
      street?: string;
      building?: string;
      phone?: string;
      email?: string;
      homepageUrl?: string;
      note?: string;
    };
    protobufAny: {
      "@type"?: string;
    } & {
      [key: string]: unknown;
    };
    rpcStatus: {
      /** Format: int32 */
      code?: number;
      message?: string;
      details?: components["schemas"]["protobufAny"][];
    };
    v1GetProfileResponse: {
      profile?: components["schemas"]["v1Profile"];
    };
    v1Profile: {
      tenantId?: string;
      postalCode?: string;
      prefecture?: string;
      city?: string;
      street?: string;
      building?: string;
      phone?: string;
      email?: string;
      homepageUrl?: string;
      note?: string;
    };
    v1UpsertProfileResponse: {
      profile?: components["schemas"]["v1Profile"];
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
  ProfileService_GetProfile: {
    parameters: {
      query?: never;
      header?: never;
      path: {
        tenantId: string;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description A successful response. */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["v1GetProfileResponse"];
        };
      };
      /** @description An unexpected error response. */
      default: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["rpcStatus"];
        };
      };
    };
  };
  ProfileService_UpsertProfile: {
    parameters: {
      query?: never;
      header?: never;
      path: {
        tenantId: string;
      };
      cookie?: never;
    };
    requestBody: {
      content: {
        "application/json": components["schemas"]["ProfileServiceUpsertProfileBody"];
      };
    };
    responses: {
      /** @description A successful response. */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["v1UpsertProfileResponse"];
        };
      };
      /** @description An unexpected error response. */
      default: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["rpcStatus"];
        };
      };
    };
  };
}